Project flists

+++ src/main/java/io/github/mzmine/datamodel/features/types/annotations/SpectralLibraryMatchType.java
@@ -114,7 +114,7 @@ public class SpectralLibraryMatchType extends ModularType implements AnnotationT
       data.set(IonAdductType.class, entry.getField(DBEntryField.ION_TYPE).orElse(""));
       data.set(SmilesStructureType.class, entry.getField(DBEntryField.SMILES).orElse(""));
       data.set(InChIStructureType.class, entry.getField(DBEntryField.INCHI).orElse(""));
-      data.set(CosineScoreType.class, score.getScore());
+      data.set(CosineScoreType.class, (float) score.getScore());
       data.set(MatchingSignalsType.class, score.getOverlap());
       if (entry.getField(DBEntryField.MZ).isPresent()) {
         data.set(PrecursorMZType.class, entry.getField(DBEntryField.MZ).get());
